third_party/blink/web_tests/scrollingcoordinator/non-fast-scrollable-region-transformed-iframe-expected.txt

This test ensures non-fast scrollable areas are calculated correctly when a CSS scale transformation is applied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nonFastScrollableRects.length is 1
PASS rectToString(nonFastScrollableRects[0]) is "[146, 146, 240, 240]"
PASS successfullyParsed is true

TEST COMPLETE
